Update README.md
Realised after reading the examples on Beaglebone and Raspberry Pi, I got it wrong. Need to deploy Node Red on Tessel instead on on localhost of terminal/computer. So, doing 127.0.0.1:1880 won't work. 
It has to be on something like http://tessel.localhost:1880.  And this is started on command line like node-red-pi...at least on the examples. 

Maybe we need to look at  http://nodered.org/docs/embedding.html to figure out on Node Red side?